+++ examples/cdc/aws/kinesis_put.py
@@ -1,11 +1,11 @@
 from yarl import URL
 
-from cratedb_toolkit.io.kinesis.adapter import KinesisAdapter
+from cratedb_toolkit.io.kinesis.adapter import KinesisStreamAdapter
 from tests.io.test_processor import DYNAMODB_CDC_INSERT_NESTED, DYNAMODB_CDC_MODIFY_NESTED, wrap_kinesis
 
 
 def main():
-    ka = KinesisAdapter(URL("kinesis://LSIAQAAAAAAVNCBMPNSG:dummy@localhost:4566/cdc-stream?region=eu-central-1"))
+    ka = KinesisStreamAdapter(URL("kinesis://LSIAQAAAAAAVNCBMPNSG:dummy@localhost:4566/cdc-stream?region=eu-central-1"))
     ka.produce(wrap_kinesis(DYNAMODB_CDC_INSERT_NESTED))
     ka.produce(wrap_kinesis(DYNAMODB_CDC_MODIFY_NESTED))