+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Dependencies: Migrated from `zyp` to `tikray`. It's effectively the
   same, but provided using a dedicated package now
+- Dependencies: Updated to `croud-1.14`
 
 ## 2025/05/13 v0.0.35
 - Added lost `pytest` dependencies to `cratedb-toolkit[testing]`