+++ CHANGES.md
@@ -8,6 +8,7 @@
 - CDC: Added canonical SQL example for PostgreSQL from Ibis
 - CDC: Enabled loading DMS events from Kinesis streams and stream-dump files
 - CDC: Added subcommand `ctk dms table-mappings`
+- Dependencies: Updated to `commons-codec>=0.0.23`
 
 ## 2025/05/13 v0.0.35
 - Added lost `pytest` dependencies to `cratedb-toolkit[testing]`